+++ packages/fluentui/react-bindings/src/hooks/useStyles.ts
@@ -46,6 +46,7 @@ type InlineStyleProps<StyleProps> = {
 };
 
 const defaultContext: StylesContextValue<{ renderRule: RendererRenderRule }> = {
+  rtl: false,
   disableAnimations: false,
   performance: {
     enableSanitizeCssPlugin: process.env.NODE_ENV !== 'production',