+++ packages/react/src/Dialog.ts
@@ -1,0 +1,2 @@
+export * from 'office-ui-fabric-react/lib/Dialog';
+export { Dialog as default } from 'office-ui-fabric-react/lib/Dialog';